Added Thrust version for disordered ham lattice
@headmyshoulder
I had to use CUSPARSE library from CUDA Toolkit to implement this. CUSPARSE
does not support CPU path, so we only will be able to show GPU times for this.
This is not Thrust-only solution, but since CUSPARSE is standard for CUDA, I
think it is legitimate. Also, I think we should have it in paper for
completeness.