Changed reference line in Phase Oscillator Chain example
Switched to ViennaCL (Tesla) as a reference line: I think it makes more sense
than Thrust in this specific plot, since Thrust's performance is so bad. This
way its easier to compare other libraries. But may be we should stick to a
single point of reference (Thrust) in all of the plots, what do you think?
@headmyshoulder, @karlrupp, @pgottsch